pipelines/runner/pyproject.toml
[project]
name = "runner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1.8",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "rich>=14.1.0",
    "sh>=2.2.2",
]

[dependency-groups]
dev = [
    "ipython>=8.18.1",
]
